+++ TechChallenge.UpdateContact/Application/Dtos/Events/ContactDeletedEventDto.cs
@@ -1,4 +1,4 @@
-﻿namespace TechChallenge.CreateContact.Application.Dtos.Events;
+﻿namespace TechChallenge.UpdateContact.Application.Dtos.Events;
 
 public class ContactDeletedEventDto
 {
