+++ TechChallenge.UpdateContact/Application/Interfaces/IContactService.cs
@@ -1,8 +1,8 @@
-﻿using TechChallenge.CreateContact.Application.Dtos;
-using TechChallenge.CreateContact.Application.Dtos.Events;
+﻿using TechChallenge.UpdateContact.Application.Dtos;
+using TechChallenge.UpdateContact.Application.Dtos.Events;
 using TechChallenge.UpdateContact.Application.Dtos;
 
-namespace TechChallenge.CreateContact.Application.Interfaces;
+namespace TechChallenge.UpdateContact.Application.Interfaces;
 
 public interface IContactService
 {
