+++ TechChallenge.UpdateContact/Application/Profiles/ContactProfile.cs
@@ -1,19 +1,19 @@
 ﻿using AutoMapper;
-using TechChallenge.CreateContact.Domain.Entities;
+using TechChallenge.UpdateContact.Domain.Entities;
 using TechChallenge.UpdateContact.Application.Dtos;
 
-namespace TechChallenge.CreateContact.Application.Profiles;
+namespace TechChallenge.UpdateContact.Application.Profiles;
 
 public class ContactProfile : Profile
 {
     public ContactProfile()
     {
-        CreateContactMapping();
+        UpdateContactMapping();
     }
 
-    private void CreateContactMapping()
+    private void UpdateContactMapping()
     {
-        CreateMap<CreateContactDto, Contact>()
+        CreateMap<UpdateContactDto, Contact>()
             .ForMember(dest => dest.PhoneAreaCode, options => options.MapFrom(src => src.Phone.AreaCode))
             .ForMember(dest => dest.Phone, options => options.MapFrom(src => src.Phone.Number));
     }
