+++ TechChallenge.UpdateContact/Domain/Entities/EntityBase.cs
@@ -1,4 +1,4 @@
-﻿namespace TechChallenge.CreateContact.Domain.Entities;
+﻿namespace TechChallenge.UpdateContact.Domain.Entities;
 public class EntityBase
 {
 }
