+++ TechChallenge.UpdateContact/Domain/Entities/PhoneArea.cs
@@ -1,4 +1,4 @@
-﻿namespace TechChallenge.CreateContact.Domain.Entities;
+﻿namespace TechChallenge.UpdateContact.Domain.Entities;
 
 public class PhoneArea: EntityBase
 {
