+++ TechChallenge.UpdateContact/Domain/Exceptions/ExchangeNotFoundException.cs
@@ -1,4 +1,4 @@
-﻿namespace TechChallenge.CreateContact.Domain.Exceptions;
+﻿namespace TechChallenge.UpdateContact.Domain.Exceptions;
 
 public class ExchangeNotFoundException(string publisherKey): Exception($"No exchange configured for publisher key: {publisherKey}")
 {
