--- TechChallenge.UpdateContact/Infrastructure/Database/CreateContactDbContext.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace TechChallenge.CreateContact.Infrastructure.Database;
-
-public class CreateContactDbContext(DbContextOptions dbContextOptions) : DbContext(dbContextOptions)
-{
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-    {
-        if (!optionsBuilder.IsConfigured)
-            optionsBuilder.UseNpgsql("Host=127.0.0.1;Port=5432;Database=create-contact;Username=postgres;Password=postgres");
-    }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        modelBuilder.ApplyConfigurationsFromAssembly(typeof(CreateContactDbContext).Assembly);
-
-        base.OnModelCreating(modelBuilder);
-    }
-}
