+++ TechChallenge.UpdateContact/Infrastructure/Services/MessageConsumerService.cs
@@ -2,9 +2,9 @@
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 using System.Text;
-using TechChallenge.CreateContact.Infrastructure.Interfaces;
+using TechChallenge.UpdateContact.Infrastructure.Interfaces;
 
-namespace TechChallenge.CreateContact.Infrastructure.Services;
+namespace TechChallenge.UpdateContact.Infrastructure.Services;
 
 public class MessageConsumerService<T>(IServiceScopeFactory serviceScopeFactory, string consumerConfigKey) : IMessageConsumerService<T>
 {
